Uniform all prefix for all commands
