+++ src/main/java/org/jboss/forge/arquillian/command/SetupWizard.java
@@ -43,7 +43,7 @@ public class SetupWizard extends AbstractProjectCommand implements UIWizard {
     @SuppressWarnings("unchecked")
     public NavigationResult next(UINavigationContext context) throws Exception {
         return Results.navigateTo(
-            AddArquillianCommand.class,
+            ArquillianAddCommand.class,
             AddTestFrameworkCommand.class);
     }
