+++ src/main/java/org/jboss/forge/arquillian/command/algeron/publisher/AlgeronAddGitPublisherCommand.java
@@ -1,6 +1,8 @@
-package org.jboss.forge.arquillian.command.algeron.retriever;
+package org.jboss.forge.arquillian.command.algeron.publisher;
 
 import org.jboss.forge.addon.dependencies.builder.DependencyBuilder;
+import org.jboss.forge.addon.facets.FacetFactory;
+import org.jboss.forge.addon.projects.ProjectFactory;
 import org.jboss.forge.addon.ui.context.UIBuilder;
 import org.jboss.forge.addon.ui.context.UIContext;
 import org.jboss.forge.addon.ui.hints.InputType;
@@ -9,17 +11,26 @@ import org.jboss.forge.addon.ui.metadata.UICommandMetadata;
 import org.jboss.forge.addon.ui.metadata.WithAttributes;
 import org.jboss.forge.addon.ui.util.Categories;
 import org.jboss.forge.addon.ui.util.Metadata;
-import org.jboss.forge.arquillian.util.URLUIValidator;
 
 import javax.inject.Inject;
 import java.util.LinkedHashMap;
 import java.util.Map;
 
-public class AddAlgeronGitRetriever extends AbstractAlgeronRetrieverCommand {
+public class AlgeronAddGitPublisherCommand extends AbstractAlgeronPublisherCommand {
+
+    @Inject
+    private ProjectFactory projectFactory;
+
+    @Inject
+    private FacetFactory facetFactory;
 
     @Inject
     @WithAttributes(shortName = 'u', label = "Git Url", required = true)
     private UIInput<String> url;
+
+    @Inject
+    @WithAttributes(shortName = 'm', label = "Comment", required = true)
+    private UIInput<String> comment;
 
     @Inject
     @WithAttributes(shortName = 's', label = "Username")
@@ -38,6 +49,10 @@ public class AddAlgeronGitRetriever extends AbstractAlgeronRetrieverCommand {
     private UIInput<String> key;
 
     @Inject
+    @WithAttributes(shortName = 'r', label = "Remote", type = InputType.FILE_PICKER)
+    private UIInput<String> remote;
+
+    @Inject
     @WithAttributes(shortName = 'o', label = "Repository")
     private UIInput<String> repository;
 
@@ -53,29 +68,39 @@ public class AddAlgeronGitRetriever extends AbstractAlgeronRetrieverCommand {
     @WithAttributes(shortName = 'b', label = "Branch")
     private UIInput<String> branch;
 
+    @Inject
+    @WithAttributes(shortName = 'e', label = "Email")
+    private UIInput<String> email;
+
+    @Inject
+    @WithAttributes(shortName = 'c', label = "Contracts Folder")
+    private UIInput<String> contractFolder;
+
     @Override
     public UICommandMetadata getMetadata(UIContext context) {
         return Metadata.from(super.getMetadata(context), getClass())
             .category(Categories.create("Algeron"))
-            .name("Arquillian Algeron: Add Git Retriever")
-            .description("This command registers a Git Retriever for Algeron");
+            .name("Arquillian Algeron: Add Git Publisher")
+            .description("This command registers a Git Publisher for Algeron");
     }
 
     @Override
     public void initializeUI(UIBuilder builder) throws Exception {
-        url.addValidator(new URLUIValidator());
-        builder.add(url)
-            .add(username)
+        builder.add(url).add(contractFolder)
+            .add(comment).add(username)
             .add(password).add(passphrase)
-            .add(key)
+            .add(key).add(remote)
             .add(repository).add(contractGitDirectory)
-            .add(tag).add(branch);
+            .add(tag).add(branch)
+            .add(email);
+        contractFolder.setDefaultValue("target/pacts");
+        remote.setDefaultValue("origin");
     }
 
     @Override
-    protected DependencyBuilder getRetrieverDependency() {
+    protected DependencyBuilder getPublisherDependency() {
         return DependencyBuilder.create().setGroupId("org.arquillian.universe")
-            .setArtifactId("arquillian-algeron-git-retriever").setPackaging("pom").setScopeType("test");
+            .setArtifactId("arquillian-algeron-git-publisher").setPackaging("pom").setScopeType("test");
     }
 
     @Override
@@ -83,6 +108,9 @@ public class AddAlgeronGitRetriever extends AbstractAlgeronRetrieverCommand {
         final Map<String, String> parameters = new LinkedHashMap<>();
         parameters.put("provider", "git");
         parameters.put("url", url.getValue());
+        parameters.put("comment", comment.getValue());
+        parameters.put("remote", remote.getValue());
+        parameters.put("contractsFolder", contractFolder.getValue());
 
         if (username.hasValue()) {
             parameters.put("username", username.getValue());
@@ -116,6 +144,10 @@ public class AddAlgeronGitRetriever extends AbstractAlgeronRetrieverCommand {
             parameters.put("branch", branch.getValue());
         }
 
+        if (email.hasValue()) {
+            parameters.put("email", email.getValue());
+        }
+
         return parameters;
     }
 
@@ -123,5 +155,4 @@ public class AddAlgeronGitRetriever extends AbstractAlgeronRetrieverCommand {
     protected String getName() {
         return "Git";
     }
-
 }
