+++ src/main/java/org/jboss/forge/arquillian/command/algeron/retriever/AlgeronAddUrlRetrieverCommand.java
@@ -1,7 +1,5 @@
-package org.jboss.forge.arquillian.command.algeron.publisher;
+package org.jboss.forge.arquillian.command.algeron.retriever;
 
-import org.jboss.forge.addon.facets.FacetFactory;
-import org.jboss.forge.addon.projects.ProjectFactory;
 import org.jboss.forge.addon.ui.context.UIBuilder;
 import org.jboss.forge.addon.ui.context.UIContext;
 import org.jboss.forge.addon.ui.input.UIInput;
@@ -9,39 +7,29 @@ import org.jboss.forge.addon.ui.metadata.UICommandMetadata;
 import org.jboss.forge.addon.ui.metadata.WithAttributes;
 import org.jboss.forge.addon.ui.util.Categories;
 import org.jboss.forge.addon.ui.util.Metadata;
+import org.jboss.forge.arquillian.util.URLUIValidator;
 
 import javax.inject.Inject;
 import java.util.LinkedHashMap;
 import java.util.Map;
 
-public class AddAlgeronUrlPublisher extends AbstractAlgeronPublisherCommand {
-
+public class AlgeronAddUrlRetrieverCommand extends AbstractAlgeronRetrieverCommand {
     @Inject
-    private ProjectFactory projectFactory;
-
-    @Inject
-    private FacetFactory facetFactory;
-
-    @Inject
-    @WithAttributes(shortName = 'u', label = "Url to POST", required = true)
+    @WithAttributes(shortName = 'u', label = "Url to GET", required = true)
     private UIInput<String> url;
-
-    @Inject
-    @WithAttributes(shortName = 'c', label = "Contracts Folder")
-    private UIInput<String> contractFolder;
 
     @Override
     public UICommandMetadata getMetadata(UIContext context) {
         return Metadata.from(super.getMetadata(context), getClass())
             .category(Categories.create("Algeron"))
-            .name("Arquillian Algeron: Add Url Publisher")
-            .description("This command registers a Url Publisher for Algeron");
+            .name("Arquillian Algeron: Add Url Retriever")
+            .description("This command registers a Url Retriever for Algeron");
     }
 
     @Override
     public void initializeUI(UIBuilder builder) throws Exception {
-        builder.add(url).add(contractFolder);
-        contractFolder.setDefaultValue("target/pacts");
+        url.addValidator(new URLUIValidator());
+        builder.add(url);
     }
 
     @Override
@@ -49,7 +37,6 @@ public class AddAlgeronUrlPublisher extends AbstractAlgeronPublisherCommand {
         Map<String, String> parameters = new LinkedHashMap<>();
         parameters.put("provider", "url");
         parameters.put("url", url.getValue());
-        parameters.put("contractsFolder", contractFolder.getValue());
         return parameters;
     }
 
@@ -57,4 +44,5 @@ public class AddAlgeronUrlPublisher extends AbstractAlgeronPublisherCommand {
     protected String getName() {
         return "Url";
     }
+
 }
